requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.9
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
jsonfield==3.1.0
pycparser==2.21
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
